+++ tests/funcs/test_pattern_rec.py
@@ -6,6 +6,8 @@
 
 # -*- coding: utf-8 -*-
 
+import os
+
 import pandas as pd
 import pytest
 
@@ -22,6 +24,34 @@ def data():
 @pytest.fixture
 def field(data):
     return data.columns[0]
+
+
+@pytest.mark.filterwarnings("ignore::DeprecationWarning")
+def test_flagPlateau():
+    path = os.path.join(
+        os.path.abspath(""), "docs/resources/data/turbidity_plateaus.csv"
+    )
+    dat = pd.read_csv(path, parse_dates=[0], index_col=0)
+    dat = dat.interpolate("linear")
+    dat = dat.ffill().bfill()
+    qc = SaQC(dat)
+    qc = qc.flagPlateau(
+        "base3", min_length="10min", max_length="7d", granularity="20min"
+    )
+    anomalies = [
+        (0, 0),
+        (5313, 5540),
+        (10000, 10200),
+        (15000, 15500),
+        (17000, 17114),
+        (17790, 17810),
+    ]
+    f = qc["base3"].flags.to_pandas().squeeze() > 0
+    for i in range(1, len(anomalies)):
+        a_slice = slice(anomalies[i][0], anomalies[i][1])
+        na_slice = slice(anomalies[i - 1][1], anomalies[i][0])
+        assert f.iloc[a_slice].all()
+        assert not (f.iloc[na_slice].any())
 
 
 @pytest.mark.parametrize("plot", [True, False])